+++ CHANGES.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
  [v0.4.0]  - unreleased
 ---------------------------------
 * Added support for audio level feature for videoroom
-
+* Added support for h265, av1 codec support for videoroom
 
  [v0.3.0]  - 2020-06-21
 ---------------------------------